+++ src/Instance/HoldingsList/Holding/index.js
@@ -1,0 +1,2 @@
+export { default as Holding } from './Holding';
+export { default as HoldingContainer } from './HoldingContainer';