+++ test/bigtest/interactors/item-create-page.js
@@ -2,6 +2,7 @@ import {
   interactor,
   clickable,
   text,
+  isPresent,
 } from '@bigtest/interactor';
 
 @interactor class HeaderDropdown {
@@ -13,9 +14,15 @@ import {
 }
 
 @interactor class ItemCreatePage {
+  isLoaded = isPresent('[data-test-header-title]');
+
   title = text('[data-test-header-title]');
   headerDropdown = new HeaderDropdown('[data-pane-header-actions-dropdown]');
   headerDropdownMenu = new HeaderDropdownMenu();
+
+  whenLoaded() {
+    return this.timeout(6000).when(() => this.isLoaded);
+  }
 }
 
 export default new ItemCreatePage('[data-test-item-page-type="create"]');